+++ sdk/resourcemanager/batch/armbatch/zz_generated_certificate_client.go
@@ -61,7 +61,7 @@ func NewCertificateClient(subscriptionID string, credential azcore.TokenCredenti
 // you do not need to run this operation after the deletion failed. You must make
 // sure that the certificate is not being used by any resources, and then you can try again to delete the certificate.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 // resourceGroupName - The name of the resource group that contains the Batch account.
 // accountName - The name of the Batch account.
 // certificateName - The identifier for the certificate. This must be made up of algorithm and thumbprint separated by a dash,
@@ -107,7 +107,7 @@ func (client *CertificateClient) cancelDeletionCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-01-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -127,7 +127,7 @@ func (client *CertificateClient) cancelDeletionHandleResponse(resp *http.Respons
 
 // Create - Creates a new certificate inside the specified account.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 // resourceGroupName - The name of the resource group that contains the Batch account.
 // accountName - The name of the Batch account.
 // certificateName - The identifier for the certificate. This must be made up of algorithm and thumbprint separated by a dash,
@@ -173,7 +173,7 @@ func (client *CertificateClient) createCreateRequest(ctx context.Context, resour
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-01-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	if options != nil && options.IfMatch != nil {
 		req.Raw().Header["If-Match"] = []string{*options.IfMatch}
@@ -199,7 +199,7 @@ func (client *CertificateClient) createHandleResponse(resp *http.Response) (Cert
 
 // BeginDelete - Deletes the specified certificate.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 // resourceGroupName - The name of the resource group that contains the Batch account.
 // accountName - The name of the Batch account.
 // certificateName - The identifier for the certificate. This must be made up of algorithm and thumbprint separated by a dash,
@@ -221,7 +221,7 @@ func (client *CertificateClient) BeginDelete(ctx context.Context, resourceGroupN
 
 // Delete - Deletes the specified certificate.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 func (client *CertificateClient) deleteOperation(ctx context.Context, resourceGroupName string, accountName string, certificateName string, options *CertificateClientBeginDeleteOptions) (*http.Response, error) {
 	req, err := client.deleteCreateRequest(ctx, resourceGroupName, accountName, certificateName, options)
 	if err != nil {
@@ -261,7 +261,7 @@ func (client *CertificateClient) deleteCreateRequest(ctx context.Context, resour
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-01-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -269,7 +269,7 @@ func (client *CertificateClient) deleteCreateRequest(ctx context.Context, resour
 
 // Get - Gets information about the specified certificate.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 // resourceGroupName - The name of the resource group that contains the Batch account.
 // accountName - The name of the Batch account.
 // certificateName - The identifier for the certificate. This must be made up of algorithm and thumbprint separated by a dash,
@@ -314,7 +314,7 @@ func (client *CertificateClient) getCreateRequest(ctx context.Context, resourceG
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-01-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -334,7 +334,7 @@ func (client *CertificateClient) getHandleResponse(resp *http.Response) (Certifi
 
 // NewListByBatchAccountPager - Lists all of the certificates in the specified account.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 // resourceGroupName - The name of the resource group that contains the Batch account.
 // accountName - The name of the Batch account.
 // options - CertificateClientListByBatchAccountOptions contains the optional parameters for the CertificateClient.ListByBatchAccount
@@ -396,7 +396,7 @@ func (client *CertificateClient) listByBatchAccountCreateRequest(ctx context.Con
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
-	reqQP.Set("api-version", "2022-01-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -413,7 +413,7 @@ func (client *CertificateClient) listByBatchAccountHandleResponse(resp *http.Res
 
 // Update - Updates the properties of an existing certificate.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-01-01
+// Generated from API version 2022-06-01
 // resourceGroupName - The name of the resource group that contains the Batch account.
 // accountName - The name of the Batch account.
 // certificateName - The identifier for the certificate. This must be made up of algorithm and thumbprint separated by a dash,
@@ -459,7 +459,7 @@ func (client *CertificateClient) updateCreateRequest(ctx context.Context, resour
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-01-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	if options != nil && options.IfMatch != nil {
 		req.Raw().Header["If-Match"] = []string{*options.IfMatch}